# packages/shared/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Norbert's Spark API",
    "description": "API for Norbert's Spark AI CRM application",
    "version": "1.0.0",
    "contact": {
      "name": "Norbert's Spark API",
      "email": "api@example.com"
    },
    "license": {
      "name": "AGPLv3",
      "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api/v1",
      "description": "Development server (HTTP) - API v1"
    },
    {
      "url": "https://127.0.0.1:3001/api/v1",
      "description": "Development server (HTTPS) - API v1"
    },
    {
      "url": "http://localhost:3001",
      "description": "Development server (HTTP) - Root"
    },
    {
      "url": "https://127.0.0.1:3001",
      "description": "Development server (HTTPS) - Root"
    }
  ],
...
  "paths": {
...
    "/ai/chat": {
      "post": {
        "summary": "Get AI-generated response",
        "description": "Sends a prompt to the AI service and retrieves the generated response",
        "operationId": "getAIResponse",
        "tags": ["ai"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "AI response stream using Server-Sent Events (SSE)",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "format": "event-stream",
                  "description": "Server-Sent Events (SSE) stream following the AI SDK stream protocol. Each line starts with 'data: ' followed by a JSON object. The stream includes various event types: start, text-delta (incremental text chunks), tool calls, finish, etc. Stream ends with 'data: [DONE]'. Full protocol: https://ai-sdk.dev/docs/ai-sdk-ui/stream-protocol"
                },
                "examples": {
                  "simpleTextResponse": {
                    "summary": "Simple text response",
                    "description": "Basic streaming response with text only",
                    "value": "data: {\"type\":\"start\"}\n\ndata: {\"type\":\"start-step\"}\n\ndata: {\"type\":\"text-start\",\"id\":\"0\"}\n\ndata: {\"type\":\"text-delta\",\"id\":\"0\",\"delta\":\"Hello\"}\n\ndata: {\"type\":\"text-delta\",\"id\":\"0\",\"delta\":\" world\"}\n\ndata: {\"type\":\"text-delta\",\"id\":\"0\",\"delta\":\"!\"}\n\ndata: {\"type\":\"text-end\",\"id\":\"0\"}\n\ndata: {\"type\":\"finish-step\"}\n\ndata: {\"type\":\"finish\",\"finishReason\":\"stop\"}\n\ndata: [DONE]"
                  },
                  "responseWithToolCall": {
                    "summary": "Response with tool invocation",
                    "description": "Streaming response that includes an AI tool call and subsequent response",
                    "value": "data: {\"type\":\"start\"}\n\ndata: {\"type\":\"start-step\"}\n\ndata: {\"type\":\"text-start\",\"id\":\"0\"}\n\ndata: {\"type\":\"text-delta\",\"id\":\"0\",\"delta\":\"Let me check that for you\"}\n\ndata: {\"type\":\"text-end\",\"id\":\"0\"}\n\ndata: {\"type\":\"tool-input-start\",\"toolCallId\":\"abc123\",\"toolName\":\"myTool\"}\n\ndata: {\"type\":\"tool-input-delta\",\"toolCallId\":\"abc123\",\"inputTextDelta\":\"{\\\"param\\\":\\\"value\\\"}\"}\n\ndata: {\"type\":\"tool-input-available\",\"toolCallId\":\"abc123\",\"toolName\":\"myTool\",\"input\":{\"param\":\"value\"}}\n\ndata: {\"type\":\"tool-output-available\",\"toolCallId\":\"abc123\",\"output\":{\"result\":\"data\"}}\n\ndata: {\"type\":\"finish-step\"}\n\ndata: {\"type\":\"start-step\"}\n\ndata: {\"type\":\"text-start\",\"id\":\"1\"}\n\ndata: {\"type\":\"text-delta\",\"id\":\"1\",\"delta\":\"Based on the result...\"}\n\ndata: {\"type\":\"text-end\",\"id\":\"1\"}\n\ndata: {\"type\":\"finish-step\"}\n\ndata: {\"type\":\"finish\",\"finishReason\":\"stop\"}\n\ndata: [DONE]"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "missingFields": {
                    "summary": "Missing required fields",
                    "value": {
                      "success": false,
                      "error": "Invalid request body",
                      "details": "id and trigger are required"
                    }
                  },
                  "invalidIdFormat": {
                    "summary": "Invalid chat ID format",
                    "value": {
                      "success": false,
                      "error": "Invalid id format",
                      "details": "incorrect ChatId format"
                    }
                  },
                  "invalidMessages": {
                    "summary": "Invalid messages format",
                    "value": {
                      "success": false,
                      "error": "Invalid request body",
                      "details": "validation error message"
                    }
                  },
                  "noMessages": {
                    "summary": "No messages provided",
                    "value": {
                      "success": false,
                      "error": "No messages provided"
                    }
                  },
                  "lastMessageNotUser": {
                    "summary": "Last message not from user",
                    "value": {
                      "success": false,
                      "error": "Last message must be from the user"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid authentication token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "success": false,
                  "error": "User not authenticated"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "success": false,
                  "error": "AI service configuration error"
                }
              }
            }
          }
        }
      }
    },
    "/ai/chats/{userId}": {
      "get": {
        "summary": "Get chat IDs for user",
        "description": "Retrieves a list of chat IDs associated with the specified user",
        "operationId": "getAIChatsByUserId",
        "tags": ["ai"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "description": "The user's unique identifier whose chat history is being retrieved",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "AI response found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIUserIdResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "invalidUserId": {
                    "summary": "Invalid userId parameter",
                    "value": {
                      "success": false,
                      "error": "Invalid userId parameter"
                    }
                  },
                  "invalidFormat": {
                    "summary": "Invalid userId format",
                    "value": {
                      "success": false,
                      "error": "Invalid userId format"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "success": false,
                  "error": "Authentication required"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User does not have required role (user, moderator or admin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "success": false,
                  "error": "Access denied. You can only access your own chat history or must have admin/moderator role"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "success": false,
                  "error": "Internal server error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AIUserIdResponse": {
        "type": "object",
        "description": "Structured response containing array of chat IDs for a user",
        "required": ["success", "data"],
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Indicates if the request was successful",
            "example": true
          },
          "data": {
            "type": "array",
            "description": "Array of UUID strings for previous AI chats",
            "items": {
              "type": "string",
              "format": "uuid",
              "example": "019b659a-2ad2-7fd8-9f32-35624caef900"
            }
          }
        }
      },
      "AIRequest": {
        "type": "object",
        "description": "AI-generated response object",
        "required": ["id", "messages", "trigger"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "019b659a-2ad2-7fd8-9f32-35624caef900"
          },
          "messages": {
            "type": "array",
            "items": {
              "type": "object",
              "required": ["id", "role", "parts"],
              "properties": {
                "id": {
                  "type": "string",
                  "example": "Z9HbJCTLArucKCzf"
                },
                "role": {
                  "type": "string",
                  "enum": ["user", "assistant"]
                },
                "parts": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": ["type"],
                    "properties": {
                      "type": {
                        "type": "string",
                        "enum": ["text", "step-start"]
                      },
                      "text": {
                        "type": "string"
                      },
                      "state": {
                        "type": "string",
                        "enum": ["done"]
                      }
                    }
                  }
                }
              }
            }
          },
          "trigger": {
            "type": "string",
            "example": "submit-message",
            "description": "Indicates which user or system action caused this AI response to be generated. Implementations should use consistent, machine-readable identifiers (e.g. \"submit-message\")."
          }
        }
      },
...
      "Error": {
        "type": "object",
        "description": "Standard error response object",
        "required": ["success", "error"],
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Invalid request data"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token for authenticated requests"
      }
    }
  }
}